Update Firebase configuration and project structure for The Shelf app
This commit modifies the Firebase configuration in multiple files, including firebase.json, GoogleService-Info.plist for iOS and macOS, and firebase_options.dart, reflecting a new project ID and updated API keys. Additionally, it updates the AndroidManifest.xml and build.gradle files to change the application ID and namespace, enhancing the app's structure. New files for key properties and privacy policies for both Fruity Blast and The Shelf are also added, improving compliance and user information transparency. The MainActivity and WebViewActivity are introduced to support the app's functionality.